DET-85 Move contact interaction
